feat: favorite and unfavorite
If authenticated, user is able to favorite and unfavorite a recipe on
the detail page. However, currently the user has to go back and refresh
for the detail page to update (button toggled).
Next we need to make it redirect to itself (refresh). Later we can try
AJAX to make it more responsive.